Rename and Move `TimeOnly` file
- Original filename is `lib.es2020.date.d.ts`
